+++ src/clsvar_gamsdomain.cpp
@@ -145,7 +145,7 @@ DEC_DECL_VARCLASSIFY(classifierClassify)
          if( domain == domainForClass[classid] )
          {
             classfound = true;
-            classForVar[varid] = classid;
+            classForVar[varid] = (int) classid;
             ++nvarsForClass[classid];
             break;
          }
